Fix Atom & RSS id fields and incorrect use of slug (#10)
* Fix atom id fields and incorrect use of slug

* Fix "self" link to be a real canonical URI

See <https://www.rfc-editor.org/rfc/rfc4287#page-22> and
<https://support.apple.com/en-gb/guide/news-publisher/apdc2c7520ff/icloud#apddf66cc2256854>

* Add getPermalink() to the feed's controller

This way both getFeedId and getSelfLink use the same function by default,
but changing one does not affect the other. This is more in line with the
real purpose of each function.

* Fix: use entry's 'id' field as guid for RSS feed

And rename the 'permalink' entry's field into 'link', because it is not
a permalink, as it contains the discussion's slug.
Also be explicit about the fact that RSS' entries' guid is a permalink
by setting 'isPermaLink' to ture, even if it is the default.